Added the option to realtime bake the curves and the mesh.
Added texture scaling for the mesh, so the textures will not be stretched when changing the resolution of the mesh.
Created a shader to merge two textures together.
Created a small script to drive the train on the rails.
The symmetry mode now also supports UV's and a texture can be applied.
Added a some materials for the walls and the rails.
And finally cleaned up some code with regards of collecting data from the scene.